fix all aggregation tests
